refactor event listeners, resolve #33
